requirements.txt
# DEEPaaS framework requirements
deepaas~=2.1.0
webargs~=5.5.2
flaat~=1.1.1

# External requirements for the model and API
fPDF2~=2.7.5
affine==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
async-timeout==4.0.2
attrs==23.1.0
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.26.4
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
dill==0.3.6
efficientnet-pytorch==0.7.1
et-xmlfile==1.1.0
evaluate==0.4.0
filelock==3.12.0
fiona==1.9.4.post1
folium==0.14.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
geopandas==0.13.2
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.0
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.2
lightning-utilities==0.8.0
lit==16.0.5.post0
lz4==4.3.2
markupsafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
munch==3.0.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pillow==9.5.0
pretrainedmodels==0.7.4
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.0
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
pytorch-lightning==2.0.1
pytorch-toolbelt==0.6.3
pytz==2023.3
pywavelets==1.4.1
pyyaml==6.0
qudida==0.0.4
rasterio==1.3.8
regex==2023.6.3
requests==2.31.0
responses==0.18.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
sympy==1.12
tensorboardx==2.6
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tokenizers==0.13.3
torchmetrics==0.11.4
tqdm==4.65.0
transformers==4.30.1
triton==2.0.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
xxhash==3.2.0
yarl==1.9.2

# Optional tools, you can use for quality of life
mlflow
